+++ Assets/Scripts/Player/Player.asmdef
@@ -5,7 +5,8 @@
         "GUID:42370f544dd6a5944b32d79fb01a6a5a",
         "GUID:941d7a5c210855c4994464255274c181",
         "GUID:560b04d1a97f54a4e82edc0cbbb69285",
-        "GUID:6910bb43451d7f349bdc3a9b4e719aaf"
+        "GUID:6910bb43451d7f349bdc3a9b4e719aaf",
+        "GUID:26d67a36a1a29324282c9ec60123e88f"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
